refactor #228 Notice PR 리뷰 반영
