+++ tracky-web/src/main/java/kernel360/trackyweb/notice/infrastructure/repository/NoticeDomainRepository.java
@@ -1,4 +1,4 @@
-package kernel360.trackyweb.notice.infrastructure;
+package kernel360.trackyweb.notice.infrastructure.repository;
 
 import kernel360.trackycore.core.infrastructure.repository.NoticeRepository;
