+++ tracky-web/src/main/java/kernel360/trackyweb/notice/infrastructure/repository/NoticeDomainRepositoryCustom.java
@@ -1,4 +1,4 @@
-package kernel360.trackyweb.notice.infrastructure;
+package kernel360.trackyweb.notice.infrastructure.repository;
 
 import java.util.List;
